Fix macOS build failures.

+++ src/audio/MacAudioDevice.cpp
@@ -557,7 +557,7 @@ bool MacAudioDevice::isRunning()
 
 int64_t MacAudioDevice::getLatencyInMicroseconds()
 {
-    std::shared_ptr<std::promise<int64_t>> prom = std::make_shared<std::promise<int> >();
+    std::shared_ptr<std::promise<int64_t>> prom = std::make_shared<std::promise<int64_t> >();
     auto fut = prom->get_future();
     enqueue_([&, prom]() {        
         // Total latency is based on the formula:
